Add files via upload
Added latest error screenshot when try to replace values

I went ahead with the remaining questions, and in the code you will see here and there that the code is not working Yet. maybe due to the replace value issue.

Also I had the heatmap working, but than the server stopped working and I had to reload. after that the heatmap did not work anymore :( despite using the same code.
I tried to close everything and clear cookies/cache, but did not help.

Hope you can guide me?